requirements_basic.txt
# Basic requirements for Streamlit Cloud deployment
streamlit
pandas
openpyxl
pdfkit
python-docx
num2words
jinja2
pypdf
numpy